Add files via upload
NumPy, which stands for "Numerical Python," is a fundamental Python library for scientific and numerical computing. It provides support for working with large, multi-dimensional arrays and matrices of numerical data, along with an extensive collection of mathematical functions to operate on these arrays. NumPy is a crucial library in the Python ecosystem for data analysis, machine learning, scientific research, and various other computational tasks.